Implemented combat cards
-Drag card to castField to activate its turn
- ActionCards can be added through editor list under Card component to be initiated on turn
- ActionCards can be clicked to cast.
- Drag cards off field to end turn (temporary resolution before completing turn-based rotation)

+++ Assets/Scripts/Card Node.cs
@@ -6,7 +6,7 @@ using UnityEngine;
 public class CardNode : ScriptableObject
 {
     public int _cardName;
-    public float _cardHP;
+    public int _cardHP;
     public int _cardATK;
     public int _cardSPD;
 }
